Refactor + tests green
