+++ nannyml/drift/ranking.py
@@ -11,7 +11,7 @@ import pandas as pd
 
 from nannyml.drift.base import DriftResult
 from nannyml.exceptions import InvalidArgumentsException
-from nannyml.metadata import ModelMetadata
+from nannyml.metadata.base import ModelMetadata
 
 
 class Ranking(abc.ABC):
@@ -56,35 +56,35 @@ class AlertCountRanking(Ranking):
     ) -> pd.DataFrame:
         """Compares the number of alerts for each feature and uses that for ranking.
 
-        Parameters
-        ----------
-        drift_calculation_result : pd.DataFrame
-            The drift calculation results. Requires alert columns to be present. These are recognized and parsed
-            using the ALERT_COLUMN_SUFFIX pattern, currently equal to ``'_alert'``.
-        model_metadata: ModelMetadata
-            Metadata describing the monitored model, used to check what the features are and exclude predictions
-            from ranking results.
-        only_drifting : bool
-            Omits features without alerts from the ranking results.
+                Parameters
+                ----------
+                drift_calculation_result : pd.DataFrame
+                    The drift calculation results. Requires alert columns to be present. These are recognized and parsed
+                    using the ALERT_COLUMN_SUFFIX pattern, currently equal to ``'_alert'``.
+                model_metadata: ModelMetadata
+                    Metadata describing the monitored model, used to check what the features are and exclude predictions
+                    from ranking results.
+                only_drifting : bool
+                    Omits features without alerts from the ranking results.
 
-        Returns
-        -------
-        feature_ranking: pd.DataFrame
-            A DataFrame containing the feature names and their ranks (the highest rank starts at 1,
-            second-highest rank is 2, etc.)
+                Returns
+                -------
+                feature_ranking: pd.DataFrame
+                    A DataFrame containing the feature names and their ranks (the highest rank starts at 1,
+                    second-highest rank is 2, etc.)
 
-        Examples
-        --------
-        >>> import nannyml as nml
-        >>> reference_df, analysis_df, target_df = nml.load_synthetic_sample()
-        >>> metadata = nml.extract_metadata(reference_df)
-        >>> metadata.target_column_name = 'work_home_actual'
-        >>> calc = nml.UnivariateStatisticalDriftCalculator(metadata, chunk_size=5000)
-        >>> calc.fit(reference_df)
-        >>> drift = calc.calculate(analysis_df)
-        >>>
-        >>> ranked = Ranker.by('alert_count').rank(drift, metadata)
-        >>> ranked
+                Examples
+                --------
+        import nannyml.metadata.extraction        >>> import nannyml as nml
+                >>> reference_df, analysis_df, target_df = nml.load_synthetic_sample()
+                >>> metadata = nannyml.metadata.extraction.extract_metadata(reference_df)
+                >>> metadata.target_column_name = 'work_home_actual'
+                >>> calc = nml.UnivariateStatisticalDriftCalculator(metadata, chunk_size=5000)
+                >>> calc.fit(reference_df)
+                >>> drift = calc.calculate(analysis_df)
+                >>>
+                >>> ranked = Ranker.by('alert_count').rank(drift, metadata)
+                >>> ranked
 
 
         """
