+++ nannyml/metadata/__init__.py
@@ -1,0 +1,9 @@
+#  Author:   Niels Nuyttens  <niels@nannyml.com>
+#
+#  License: Apache Software License 2.0
+
+from .base import ModelMetadata, ModelType
+from .binary_classification import BinaryClassificationMetadata
+from .extraction import extract_metadata
+from .feature import Feature, FeatureType
+from .multiclass_classification import MultiClassClassificationMetadata
