+++ nannyml/metadata/base.py
@@ -3,6 +3,7 @@
 #  License: Apache Software License 2.0
 
 """NannyML module providing classes and utilities for dealing with model metadata."""
+import abc
 import logging
 from enum import Enum
 from typing import Any, Dict, List, Optional, Tuple
@@ -11,6 +12,7 @@ import numpy as np
 import pandas as pd
 
 from nannyml.exceptions import MissingMetadataException
+from nannyml.metadata.feature import Feature, FeatureType
 
 NML_METADATA_PARTITION_COLUMN_NAME = 'nml_meta_partition'
 NML_METADATA_PREDICTION_COLUMN_NAME = 'nml_meta_prediction'
@@ -32,114 +34,20 @@ NML_METADATA_COLUMNS = [
 logger = logging.getLogger(__name__)
 
 
-# TODO wording
-class FeatureType(str, Enum):
-    """An enum indicating what kind of variable a given feature represents.
+class ModelType(str, Enum):
+    """A listing of all possible model types.
 
-    The FeatureType enum is a property of a Feature. NannyML uses this information to select the best drift detection
-    algorithms for each individual feature.
+    The model type will determine which specific metadata properties are required by NannyML.
+    Each :class:`~nannyml.metadata.metadata.ModelMetadata` subclass will be associated with a specific ModelType.
 
-    We consider the following feature types:
-
-    CONTINUOUS: numeric variables that have an infinite number of values between any two values.
-    CATEGORICAL: has two or more categories, but there is no intrinsic ordering to the categories.
-    ORDINAL: similar to a categorical variable, but there is a clear ordering of the categories.
-    UNKNOWN: indicates NannyML couldn't detect the feature type with a high enough degree of certainty.
     """
 
-    CONTINUOUS = 'continuous'
-    CATEGORICAL = 'categorical'
-    ORDINAL = 'ordinal'
-    UNKNOWN = 'unknown'
-
-
-class Feature:
-    """Representation of a model feature.
-
-    NannyML requires both model inputs and outputs to perform drift calculation and performance metrics.
-    It needs to understand what features a model is made of and what kind of data they might contain.
-    The Feature class allows you to provide this information.
-    """
-
-    def __init__(self, column_name: str, label: str, feature_type: FeatureType, description: str = None):
-        """Creates a new Feature instance.
-
-        The ModelMetadata class contains a list of Features that describe the values that serve as model input.
-
-        Parameters
-        ----------
-        column_name : str
-            The name of the column where the feature is found in the (to be provided) model input/output data.
-        label : str
-            A (human-friendly) label for the feature.
-        feature_type : FeatureType
-            The kind of values the data for this feature are.
-        description : str
-            Some additional information to display within results and visualizations.
-
-        Returns
-        -------
-        feature: Feature
-
-        Examples
-        --------
-        >>> from nannyml.metadata import Feature, FeatureType
-        >>> feature = Feature(column_name='dist_from_office', label='office_distance',
-        description='Distance from home to the office', feature_type=FeatureType.CONTINUOUS)
-        >>> feature
-        Feature({'label': 'office_distance', 'column_name': 'dist_from_office', 'type': 'continuous',
-        'description': 'Distance from home to the office'})
-        """
-        self.column_name = column_name
-        self.label = label
-        self.description = description
-        self.feature_type = feature_type
-
-    def to_dict(self) -> Dict[str, Any]:
-        """Converts the feature into a Dictionary representation.
-
-        Examples
-        --------
-        >>> from nannyml.metadata import Feature, FeatureType
-        >>> feature = Feature(column_name='dist_from_office', label='office_distance',
-        description='Distance from home to the office', feature_type=FeatureType.CONTINUOUS)
-        >>> feature.to_dict()
-        {'label': 'office_distance',
-         'column_name': 'dist_from_office',
-         'type': 'continuous',
-         'description': 'Distance from home to the office'}
-
-        """
-        return {
-            'label': self.label,
-            'column_name': self.column_name,
-            'type': self.feature_type.value,
-            'description': self.description,
-        }
-
-    def __repr__(self):
-        """String representation of a single Feature."""
-        return f'Feature({self.to_dict()})'
-
-    def __str__(self):  # pragma: no cover
-        """String representation of a single Feature."""
-        return f'Feature({self.to_dict()})'
-
-    def print(self):
-        """String representation of a single Feature."""
-        strs = [
-            f"Feature: {self.label}",
-            '',
-            f"{'Column name':25} {self.column_name:25}",
-            f"{'Description':25} {self.description:25}",
-            f"{'Type':25} {self.feature_type:25}",
-            '',
-        ]
-        return str.join('\n', strs)
+    CLASSIFICATION_BINARY = 'classification_binary'
+    CLASSIFICATION_MULTICLASS = 'classification_multiclass'
 
 
 # TODO wording
-class ModelMetadata:
+class ModelMetadata(abc.ABC):
     """The ModelMetadata class contains all the information nannyML requires.
 
     To understand the model inputs and outputs you wish it to process, nannyML needs to understand what your model
@@ -167,8 +75,8 @@ class ModelMetadata:
     # TODO wording
     def __init__(
         self,
+        model_type: ModelType,
         model_name: str = None,
-        model_problem: str = 'binary_classification',
         features: List[Feature] = None,
         prediction_column_name: str = None,
         predicted_probability_column_name: str = None,
@@ -180,10 +88,11 @@ class ModelMetadata:
 
         Parameters
         ----------
-        model_name : string
-            A human-readable name for the model. Required.
-        model_problem : string
-            The kind of problem your model is trying to solve. Optional, defaults to `binary_classification`.
+        model_type : ModelType
+            The kind of problem your model is trying to solve. Used to determine which metadata properties should
+            be known by NannyML.
+        model_name : string, default=None
+            A human-readable name for the model.
         features : List[Feature]
             The list of Features for the model. Optional, defaults to `None`.
         prediction_column_name : string
@@ -208,7 +117,7 @@ class ModelMetadata:
         Examples
         --------
         >>> from nannyml.metadata import ModelMetadata, Feature, FeatureType
-        >>> metadata = ModelMetadata('work_from_home', target_column_name='work_home_actual')
+        >>> metadata = ModelMetadata(model_type=ModelType.CLASSIFICATION_BINARY, target_column_name='work_home_actual')
         >>> metadata.features = [Feature(column_name='dist_from_office', label='office_distance',
         description='Distance from home to the office', feature_type=FeatureType.CONTINUOUS),
         >>> Feature(column_name='salary_range', label='salary_range', feature_type=FeatureType.CATEGORICAL)]
@@ -228,10 +137,8 @@ class ModelMetadata:
            'description': None}]}
 
         """
-        self.id: int
-
+        self.model_type = model_type
         self.name = model_name
-        self.model_problem = model_problem
 
         self._prediction_column_name = prediction_column_name
         self._predicted_probability_column_name = predicted_probability_column_name
@@ -303,7 +210,7 @@ class ModelMetadata:
         Examples
         --------
         >>> from nannyml.metadata import ModelMetadata, Feature, FeatureType
-        >>> metadata = ModelMetadata('work_from_home', target_column_name='work_home_actual')
+        >>> metadata = ModelMetadata(model_type=ModelType.CLASSIFICATION_BINARY, target_column_name='work_home_actual')
         >>> metadata.features = [Feature(column_name='dist_from_office', label='office_distance',
         description='Distance from home to the office', feature_type=FeatureType.CONTINUOUS),
         >>> Feature(column_name='salary_range', label='salary_range', feature_type=FeatureType.CATEGORICAL)]
@@ -352,7 +259,7 @@ class ModelMetadata:
         Examples
         --------
         >>> from nannyml.metadata import ModelMetadata, Feature, FeatureType
-        >>> metadata = ModelMetadata('work_from_home', target_column_name='work_home_actual')
+        >>> metadata = ModelMetadata(model_type=ModelType.CLASSIFICATION_BINARY, target_column_name='work_home_actual')
         >>> metadata.features = [Feature(column_name='dist_from_office', label='office_distance',
         >>> description='Distance to the office', feature_type=FeatureType.CONTINUOUS),
         >>> Feature(column_name='salary_range', label='salary_range', feature_type=FeatureType.CATEGORICAL)]
@@ -430,7 +337,7 @@ class ModelMetadata:
         Examples
         --------
         >>> from nannyml.metadata import ModelMetadata, Feature, FeatureType
-        >>> metadata = ModelMetadata('work_from_home', target_column_name='work_home_actual')
+        >>> metadata = ModelMetadata(model_type=ModelType.CLASSIFICATION_BINARY, target_column_name='work_home_actual')
         >>> metadata.features = [Feature(column_name='dist_from_office', label='office_distance',
         >>> description='Distance from home to the office', feature_type=FeatureType.CONTINUOUS),
         >>> Feature(column_name='salary_range', label='salary_range', feature_type=FeatureType.CATEGORICAL)]
@@ -502,7 +409,7 @@ class ModelMetadata:
         Examples
         --------
         >>> from nannyml.metadata import ModelMetadata, Feature, FeatureType
-        >>> metadata = ModelMetadata('work_from_home', target_column_name='work_home_actual')
+        >>> metadata = ModelMetadata(model_type=ModelType.CLASSIFICATION_BINARY, target_column_name='work_home_actual')
         >>> metadata.features = [
         >>>     Feature('cat1', 'cat1', FeatureType.CATEGORICAL), Feature('cat2', 'cat2', FeatureType.CATEGORICAL),
         >>>     Feature('cont1', 'cont1', FeatureType.CONTINUOUS), Feature('cont2', 'cont2', FeatureType.CONTINUOUS)]
@@ -534,6 +441,7 @@ class ModelMetadata:
         """
         return [f for f in self.features if f.feature_type == FeatureType.CONTINUOUS]
 
+    @abc.abstractmethod
     def is_complete(self) -> Tuple[bool, List[str]]:
         """Flags if the ModelMetadata is considered complete or still missing values.
 
@@ -585,132 +493,51 @@ class ModelMetadata:
 
         return complete, missing
 
+    @abc.abstractmethod
+    def extract(self, data: pd.DataFrame, model_name: str = None, exclude_columns: List[str] = None):
+        def check_for_nan(column_names):
+            number_of_nan = data[column_names].isnull().sum().sum()
+            if number_of_nan > 0:
+                raise MissingMetadataException(
+                    f'found {number_of_nan} NaN values in one of these columns: {column_names}'
+                )
+
+        if len(data.columns) == 0:
+            return None
+
+        if exclude_columns is not None and len(exclude_columns) != 0:
+            data = data.drop(columns=exclude_columns)
+
+        predictions = _guess_predictions(data)
+        check_for_nan(predictions)
+        self.prediction_column_name = None if len(predictions) == 0 else predictions[0]  # type: ignore
+
+        predicted_probabilities = _guess_predicted_probabilities(data)
+        check_for_nan(predicted_probabilities)
+        self.predicted_probability_column_name = (
+            None if len(predicted_probabilities) == 0 else predicted_probabilities[0]
+        )
+
+        targets = _guess_targets(data)
+        check_for_nan(targets)
+        self.target_column_name = None if len(targets) == 0 else targets[0]  # type: ignore
+
+        partitions = _guess_partitions(data)
+        check_for_nan(partitions)
+        self.partition_column_name = None if len(partitions) == 0 else partitions[0]  # type: ignore
+
+        timestamps = _guess_timestamps(data)
+        check_for_nan(timestamps)
+        self.timestamp_column_name = None if len(timestamps) == 0 else timestamps[0]  # type: ignore
+
+        self.features = _extract_features(data)
+
+        return self
+
     def __remove_from_features(self, column_name: str):
         current_feature = self.feature(column=column_name)
         if current_feature:
             self.features.remove(current_feature)
-
-
-def extract_metadata(data: pd.DataFrame, model_name: str = None, exclude_columns: List[str] = None):
-    """Tries to extract model metadata from a given data set.
-
-    Manually constructing model metadata can be cumbersome, especially if you have hundreds of features.
-    NannyML includes this helper function that tries to do the boring stuff for you using some simple rules.
-
-    By default, all columns in the given dataset are considered to be either model features or metadata. Use the
-    ``exclude_columns`` parameter to prevent columns from being interpreted as metadata or features.
-
-    Parameters
-    ----------
-    data : DataFrame
-        The dataset containing model inputs and outputs, enriched with the required metadata.
-    model_name : str
-        A human-readable name for the model.
-    exclude_columns: List[str], default=None
-        A list of column names that are to be skipped during metadata extraction, preventing them from being interpreted
-        as either model metadata or model features.
-
-    Returns
-    -------
-    metadata: ModelMetadata
-        A fully initialized ModelMetadata instance.
-
-
-    Examples
-    --------
-    >>> from nannyml.datasets import load_synthetic_sample
-    >>> from nannyml.metadata import extract_metadata
-    >>> ref_df, _, _ = load_synthetic_sample()
-    >>> metadata = extract_metadata(ref_df, model_name='work_from_home')
-    >>> metadata.is_complete()
-    (False, ['target_column_name'])
-    >>> metadata.to_dict()
-    {'identifier_column_name': 'identifier',
-     'timestamp_column_name': 'timestamp',
-     'partition_column_name': 'partition',
-     'target_column_name': None,
-     'prediction_column_name': None,
-     'predicted_probability_column_name': 'y_pred_proba',
-     'features': [{'label': 'distance_from_office',
-       'column_name': 'distance_from_office',
-       'type': 'continuous',
-       'description': 'extracted feature: distance_from_office'},
-      {'label': 'salary_range',
-       'column_name': 'salary_range',
-       'type': 'categorical',
-       'description': 'extracted feature: salary_range'},
-      {'label': 'gas_price_per_litre',
-       'column_name': 'gas_price_per_litre',
-       'type': 'continuous',
-       'description': 'extracted feature: gas_price_per_litre'},
-      {'label': 'public_transportation_cost',
-       'column_name': 'public_transportation_cost',
-       'type': 'continuous',
-       'description': 'extracted feature: public_transportation_cost'},
-      {'label': 'wfh_prev_workday',
-       'column_name': 'wfh_prev_workday',
-       'type': 'categorical',
-       'description': 'extracted feature: wfh_prev_workday'},
-      {'label': 'workday',
-       'column_name': 'workday',
-       'type': 'categorical',
-       'description': 'extracted feature: workday'},
-      {'label': 'tenure',
-       'column_name': 'tenure',
-       'type': 'continuous',
-       'description': 'extracted feature: tenure'},
-      {'label': 'work_home_actual',
-       'column_name': 'work_home_actual',
-       'type': 'categorical',
-       'description': 'extracted feature: work_home_actual'}]}
-
-
-    Notes
-    -----
-    NannyML can only make educated guesses as to what kind of data lives where. When NannyML feels to unsure
-    about a guess, it will not use it.
-    Be sure to always review the results of this method for their correctness and completeness.
-    Adjust and complete as you see fit.
-    """
-
-    def check_for_nan(column_names):
-        number_of_nan = data[column_names].isnull().sum().sum()
-        if number_of_nan > 0:
-            raise MissingMetadataException(f'found {number_of_nan} NaN values in one of these columns: {column_names}')
-
-    if len(data.columns) == 0:
-        return None
-
-    if exclude_columns is not None and len(exclude_columns) != 0:
-        data = data.drop(columns=exclude_columns)
-
-    metadata = ModelMetadata(model_name=model_name)
-
-    predictions = _guess_predictions(data)
-    check_for_nan(predictions)
-    metadata.prediction_column_name = None if len(predictions) == 0 else predictions[0]  # type: ignore
-
-    predicted_probabilities = _guess_predicted_probabilities(data)
-    check_for_nan(predicted_probabilities)
-    metadata.predicted_probability_column_name = (
-        None if len(predicted_probabilities) == 0 else predicted_probabilities[0]
-    )
-
-    targets = _guess_targets(data)
-    check_for_nan(targets)
-    metadata.target_column_name = None if len(targets) == 0 else targets[0]  # type: ignore
-
-    partitions = _guess_partitions(data)
-    check_for_nan(partitions)
-    metadata.partition_column_name = None if len(partitions) == 0 else partitions[0]  # type: ignore
-
-    timestamps = _guess_timestamps(data)
-    check_for_nan(timestamps)
-    metadata.timestamp_column_name = None if len(timestamps) == 0 else timestamps[0]  # type: ignore
-
-    metadata.features = _extract_features(data)
-
-    return metadata
 
 
 def _guess_timestamps(data: pd.DataFrame) -> List[str]:
