+++ tests/performance_calculation/test_performance_calculator.py
@@ -11,7 +11,8 @@ import pytest
 
 from nannyml.datasets import load_synthetic_sample
 from nannyml.exceptions import InvalidArgumentsException
-from nannyml.metadata import ModelMetadata, extract_metadata
+from nannyml.metadata import extract_metadata
+from nannyml.metadata.base import ModelMetadata, ModelType
 from nannyml.performance_calculation import PerformanceCalculator
 from nannyml.performance_calculation.metrics import AUROC, F1
 
@@ -27,7 +28,7 @@ def data() -> Tuple[pd.DataFrame, pd.DataFrame, pd.DataFrame]:  # noqa: D103
 
 @pytest.fixture
 def metadata(data) -> ModelMetadata:  # noqa: D103
-    md = extract_metadata(data[0])
+    md = extract_metadata(data[0], model_type=ModelType.CLASSIFICATION_BINARY)
     md.target_column_name = 'work_home_actual'
     return md
