+++ tests/test_metadata.py
@@ -11,7 +11,8 @@ import pandas as pd
 import pytest
 
 from nannyml.exceptions import MissingMetadataException
-from nannyml.metadata import (
+from nannyml.metadata import extract_metadata
+from nannyml.metadata.base import (
     NML_METADATA_PARTITION_COLUMN_NAME,
     NML_METADATA_PREDICTED_PROBABILITY_COLUMN_NAME,
     NML_METADATA_PREDICTION_COLUMN_NAME,
@@ -20,6 +21,7 @@ from nannyml.metadata import (
     Feature,
     FeatureType,
     ModelMetadata,
+    ModelType,
     _guess_features,
     _guess_partitions,
     _guess_predicted_probabilities,
@@ -27,8 +29,9 @@ from nannyml.metadata import (
     _guess_targets,
     _guess_timestamps,
     _predict_feature_types,
-    extract_metadata,
 )
+from nannyml.metadata.binary_classification import BinaryClassificationMetadata
+from nannyml.metadata.multiclass_classification import MultiClassClassificationMetadata
 
 
 @pytest.fixture
@@ -38,7 +41,7 @@ def sample_feature() -> Feature:  # noqa: D103
 
 @pytest.fixture
 def sample_model_metadata(sample_feature) -> ModelMetadata:  # noqa: D103
-    return ModelMetadata(model_name='my_model', features=[sample_feature])
+    return BinaryClassificationMetadata(model_name='my_model', features=[sample_feature])
 
 
 @pytest.fixture
@@ -142,10 +145,16 @@ def test_feature_string_representation_contains_all_properties(sample_feature): 
     assert 'Type' in sut
 
 
-def test_model_metadata_creation_with_defaults_has_correct_properties():  # noqa: D103
-    sut = ModelMetadata(model_name='model')
-    assert sut.name == 'model'
-    assert sut.model_problem == 'binary_classification'
+@pytest.mark.parametrize(
+    'sut,model_type',
+    [
+        (BinaryClassificationMetadata(), ModelType.CLASSIFICATION_BINARY),
+        (MultiClassClassificationMetadata(), ModelType.CLASSIFICATION_MULTICLASS),
+    ],
+)
+def test_model_metadata_creation_with_defaults_has_correct_properties(sut, model_type):  # noqa: D103
+    assert sut.name is None
+    assert sut.model_type == model_type
     assert sut.features is not None
     assert len(sut.features) == 0
     assert sut.prediction_column_name is None
@@ -156,9 +165,8 @@ def test_model_metadata_creation_with_defaults_has_correct_properties():  # noqa
 
 
 def test_model_metadata_creation_with_custom_values_has_correct_properties(sample_feature):  # noqa: D103
-    sut = ModelMetadata(
+    sut = BinaryClassificationMetadata(
         model_name='model',
-        model_problem='classification',
         features=[sample_feature],
         prediction_column_name='pred',
         predicted_probability_column_name='pred_proba',
@@ -167,7 +175,7 @@ def test_model_metadata_creation_with_custom_values_has_correct_properties(sampl
         timestamp_column_name='ts',
     )
     assert sut.name == 'model'
-    assert sut.model_problem == 'classification'
+    assert sut.model_type == ModelType.CLASSIFICATION_BINARY
     assert len(sut.features) == 1
     assert sut.features[0].label == 'label'
     assert sut.features[0].column_name == 'col'
@@ -266,24 +274,24 @@ def test_feature_filtering_without_criteria_returns_none(sample_model_metadata):
 
 
 def test_extract_metadata_for_no_cols_dataframe_should_return_none():  # noqa: D103
-    sut = extract_metadata(data=pd.DataFrame())
+    sut = extract_metadata(data=pd.DataFrame(), model_type=ModelType.CLASSIFICATION_BINARY)
     assert sut is None
 
 
 def test_extract_metadata_without_any_feature_columns_should_return_metadata_without_features():  # noqa: D103
     data = pd.DataFrame(columns=['prediction', 'actual', 'partition', 'ts'])
-    sut = extract_metadata(data)
+    sut = extract_metadata(data, model_type=ModelType.CLASSIFICATION_BINARY)
     assert len(sut.features) == 0
 
 
 def test_extract_metadata_should_not_consider_excluded_columns(sample_model_metadata, sample_data):  # noqa: D103
-    metadata = extract_metadata(sample_data, exclude_columns=['id'])
+    metadata = extract_metadata(sample_data, exclude_columns=['id'], model_type=ModelType.CLASSIFICATION_BINARY)
     assert metadata.feature(feature='id') is None
 
 
 def test_extract_metadata_for_empty_dataframe_should_return_correct_column_names(sample_model_metadata):  # noqa: D103
     data = pd.DataFrame(columns=['y_pred', 'y_pred_proba', 'actual', 'partition', 'ts', 'feat1', 'feat2'])
-    sut = extract_metadata(data)
+    sut = extract_metadata(data, model_type=ModelType.CLASSIFICATION_BINARY)
     assert sut is not None
     assert sut.prediction_column_name == 'y_pred'
     assert sut.predicted_probability_column_name == 'y_pred_proba'
@@ -295,7 +303,7 @@ def test_extract_metadata_for_empty_dataframe_should_return_correct_column_names
 # TODO verify behaviour
 def test_extract_metadata_for_empty_dataframe_should_return_features_with_feature_type_unknown():  # noqa: D103
     data = pd.DataFrame(columns=['prediction', 'actual', 'partition', 'ts', 'feat1', 'feat2'])
-    sut = extract_metadata(data)
+    sut = extract_metadata(data, model_type=ModelType.CLASSIFICATION_BINARY)
     assert len(sut.features) == 2
     assert sut.features[0].feature_type is FeatureType.UNKNOWN
     assert sut.features[1].feature_type is FeatureType.UNKNOWN
@@ -303,7 +311,7 @@ def test_extract_metadata_for_empty_dataframe_should_return_features_with_featur
 
 def test_extract_metadata_without_matching_columns_should_set_them_to_none():  # noqa: D103
     data = pd.DataFrame(columns=['a', 'b', 'c'])
-    sut = extract_metadata(data)
+    sut = extract_metadata(data, model_type=ModelType.CLASSIFICATION_BINARY)
     assert sut.prediction_column_name is None
     assert sut.target_column_name is None
     assert sut.partition_column_name is None
@@ -312,7 +320,7 @@ def test_extract_metadata_without_matching_columns_should_set_them_to_none():  #
 
 def test_extract_metadata_without_matching_columns_should_set_features():  # noqa: D103
     data = pd.DataFrame(columns=['a', 'b', 'c'])
-    sut = extract_metadata(data)
+    sut = extract_metadata(data, model_type=ModelType.CLASSIFICATION_BINARY)
     assert len(sut.features) == 3
     assert sut.feature(column='a')
     assert sut.feature(column='b')
@@ -321,7 +329,7 @@ def test_extract_metadata_without_matching_columns_should_set_features():  # noq
 
 def test_extract_metadata_with_multiple_matching_columns_should_return_first_matching_column():  # noqa: D103
     data = pd.DataFrame(columns=['target', 'ground_truth', 'actual'])
-    sut = extract_metadata(data)
+    sut = extract_metadata(data, model_type=ModelType.CLASSIFICATION_BINARY)
     assert sut.target_column_name == 'target'
 
 
@@ -329,7 +337,7 @@ def test_extract_metadata_does_not_fail_when_adding_metadata_parameter_fails(): 
     cols = ['prediction', 'actual', 'partition', 'timestamp_non_standard', 'feat1', 'feat2']
     df = pd.DataFrame(columns=cols)
     try:
-        _ = extract_metadata(df)
+        _ = extract_metadata(df, model_type=ModelType.CLASSIFICATION_BINARY)
     except Exception:
         pytest.fail("should not have failed because of inner exception")
 
@@ -338,13 +346,13 @@ def test_extract_metadata_does_not_fail_when_adding_metadata_parameter_fails(): 
 def test_extract_metadata_raises_missing_metadata_exception_when_missing_metadata_values(metadata_column):  # noqa: D103
     df = pd.DataFrame({metadata_column: [np.NaN]})
     with pytest.raises(MissingMetadataException):
-        _ = extract_metadata(df)
+        _ = extract_metadata(df, model_type=ModelType.CLASSIFICATION_BINARY)
 
 
 def test_extract_metadata_raises_missing_metadata_exception_when_missing_timestamp_values():  # noqa: D103
     df = pd.DataFrame({'timestamp': [datetime.datetime.now(), np.NaN]})
     with pytest.raises(MissingMetadataException):
-        _ = extract_metadata(df)
+        _ = extract_metadata(df, model_type=ModelType.CLASSIFICATION_BINARY)
 
 
 @pytest.mark.parametrize(
@@ -443,7 +451,7 @@ def test_feature_type_detection_sets_between_low_and_mid_cardinality_threshold_t
 
 def test_enrich_copies_each_metadata_column_to_new_fixed_column():  # noqa: D103
     data = pd.DataFrame(columns=['identity', 'prediction', 'actual', 'partition', 'ts', 'feat1', 'feat2'])
-    md = extract_metadata(data, model_name='model')
+    md = extract_metadata(data, model_name='model', model_type=ModelType.CLASSIFICATION_BINARY)
     sut = md.enrich(data).columns
 
     assert NML_METADATA_TIMESTAMP_COLUMN_NAME in sut
@@ -456,7 +464,7 @@ def test_enrich_copies_each_metadata_column_to_new_fixed_column():  # noqa: D103
 def test_enrich_works_on_copy_of_data_by_default():  # noqa: D103
     data = pd.DataFrame(columns=['identity', 'prediction', 'actual', 'partition', 'ts', 'feat1', 'feat2'])
     old_column_count = len(data.columns)
-    md = extract_metadata(data, model_name='model')
+    md = extract_metadata(data, model_name='model', model_type=ModelType.CLASSIFICATION_BINARY)
     sut = md.enrich(data).columns
 
     assert len(sut) == len(data.columns) + 5
@@ -475,7 +483,7 @@ def test_enrich_works_on_copy_of_data_by_default():  # noqa: D103
 
 
 def test_enrich_adds_nan_prediction_column_if_no_prediction_column_in_original_data(sample_data):  # noqa: D103
-    md = extract_metadata(sample_data)
+    md = extract_metadata(sample_data, model_type=ModelType.CLASSIFICATION_BINARY)
     analysis_data = sample_data.drop(columns=[md.prediction_column_name])
     sut = md.enrich(analysis_data)
     assert NML_METADATA_PREDICTION_COLUMN_NAME in sut.columns
@@ -485,7 +493,7 @@ def test_enrich_adds_nan_prediction_column_if_no_prediction_column_in_original_d
 def test_enrich_adds_nan_predicted_probability_column_if_no_predicted_probability_in_original_data(  # noqa: D103
     sample_data,
 ):
-    md = extract_metadata(sample_data)
+    md = extract_metadata(sample_data, model_type=ModelType.CLASSIFICATION_BINARY)
     analysis_data = sample_data.drop(columns=[md.prediction_column_name])
     sut = md.enrich(analysis_data)
     assert NML_METADATA_PREDICTION_COLUMN_NAME in sut.columns
@@ -493,7 +501,7 @@ def test_enrich_adds_nan_predicted_probability_column_if_no_predicted_probabilit
 
 
 def test_enrich_adds_nan_ground_truth_column_if_no_ground_truth_in_original_data(sample_data):  # noqa: D103
-    md = extract_metadata(sample_data)
+    md = extract_metadata(sample_data, model_type=ModelType.CLASSIFICATION_BINARY)
     analysis_data = sample_data.drop(columns=[md.target_column_name])
     sut = md.enrich(analysis_data)
     assert NML_METADATA_TARGET_COLUMN_NAME in sut.columns
@@ -501,7 +509,7 @@ def test_enrich_adds_nan_ground_truth_column_if_no_ground_truth_in_original_data
 
 
 def test_complete_returns_all_missing_properties_when_metadata_is_incomplete(sample_data):  # noqa: D103
-    md = extract_metadata(sample_data)
+    md = extract_metadata(sample_data, model_type=ModelType.CLASSIFICATION_BINARY)
     md.timestamp_column_name = None
     md.prediction_column_name = None
     md.predicted_probability_column_name = None
@@ -513,7 +521,7 @@ def test_complete_returns_all_missing_properties_when_metadata_is_incomplete(sam
 
 
 def test_complete_returns_complete_if_prediction_is_none_but_predicted_probabilities_are_not(sample_data):  # noqa: D103
-    md = extract_metadata(sample_data)
+    md = extract_metadata(sample_data, model_type=ModelType.CLASSIFICATION_BINARY)
     md.prediction_column_name = None
     md.predicted_probability_column_name = 'y_pred_proba'
     sut = md.is_complete()
@@ -523,7 +531,7 @@ def test_complete_returns_complete_if_prediction_is_none_but_predicted_probabili
 
 
 def test_complete_returns_complete_if_predicted_probabilities_is_none_but_prediction_is_not(sample_data):  # noqa: D103
-    md = extract_metadata(sample_data)
+    md = extract_metadata(sample_data, model_type=ModelType.CLASSIFICATION_BINARY)
     md.prediction_column_name = 'y_pred'
     md.predicted_probability_column_name = None
     sut = md.is_complete()
@@ -533,7 +541,7 @@ def test_complete_returns_complete_if_predicted_probabilities_is_none_but_predic
 
 
 def test_complete_returns_incomplete_and_both_prediction_props_when_none_of_both_set(sample_data):  # noqa: D103
-    md = extract_metadata(sample_data)
+    md = extract_metadata(sample_data, model_type=ModelType.CLASSIFICATION_BINARY)
     md.prediction_column_name = None
     md.predicted_probability_column_name = None
     sut = md.is_complete()
@@ -543,7 +551,7 @@ def test_complete_returns_incomplete_and_both_prediction_props_when_none_of_both
 
 
 def test_complete_returns_false_when_features_of_unknown_feature_type_exist(sample_data):  # noqa: D103
-    md = extract_metadata(sample_data)
+    md = extract_metadata(sample_data, model_type=ModelType.CLASSIFICATION_BINARY)
 
     # Rig the data
     md.features[0].feature_type = FeatureType.UNKNOWN
@@ -585,7 +593,7 @@ def test_continuous_features_returns_only_continuous_features(sample_model_metad
 
 def test_setting_prediction_column_name_after_extracting_metadata_updates_the_features_list(sample_data):  # noqa: D103
     sample_data.rename(columns={'output': 'prediction_score'}, inplace=True)
-    md = extract_metadata(sample_data)
+    md = extract_metadata(sample_data, model_type=ModelType.CLASSIFICATION_BINARY)
     md.prediction_column_name = 'prediction_score'
     assert md.prediction_column_name == 'prediction_score'
     assert md.feature(column='prediction_score') is None
